🐛 fix issue where the .ferret folder would automatically be deleted if nested by default

+++ packages/varmint/src/varmint-workspace-manager.ts
@@ -111,7 +111,7 @@ export const varmintWorkspaceManager = {
 			}
 			const realRootContents = fs.readdirSync(realRoot)
 			for (const rootContent of realRootContents) {
-				if (!rootMap.has(rootContent)) {
+				if (rootContent !== `.ferret` && !rootMap.has(rootContent)) {
 					const pathForRemoval = path.join(realRoot, rootContent)
 					console.log(`🧹 globalFlush: removing directory ${pathForRemoval}`)
 					fs.rmSync(pathForRemoval, { recursive: true })
